pyproject.toml
[tool.poetry]
name = "redsoft"
version = "0.1.0"
description = ""
authors = ["Борзенков Александр <alexandr.borzenkov@red-soft.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
annotated-types = "0.7.0"
anyio = "4.6.2.post1"
asyncpg = "0.30.0"
bcrypt = "4.2.0"
certifi = "2024.8.30"
charset-normalizer = "3.4.0"
click = "8.1.7"
colorama = "0.4.6"
docker = "7.1.0"
dynaconf = "3.2.6"
ecdsa = "0.19.0"
fastapi = "0.115.3"
gitdb = "4.0.11"
gitpython = "3.1.43"
h11 = "0.14.0"
idna = "3.10"
passlib = "1.7.4"
pyasn1 = "0.6.1"
pydantic = "2.9.2"
pydantic-core = "2.23.4"
python-jose = "3.3.0"
python-multipart = "0.0.12"
pywin32 = "308"
requests = "2.32.3"
rsa = "4.9"
six = "1.16.0"
smmap = "5.0.1"
sniffio = "1.3.1"
starlette = "0.41.0"
typing-extensions = "4.12.2"
urllib3 = "2.2.3"
uvicorn = "0.32.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
